+++ earth2studio/models/px/base.py
@@ -69,7 +69,6 @@ class PrognosticModel(Protocol):
         """
         pass
 
-    @property
     def input_coords(self) -> CoordSystem:
         """Input coordinate system of prognostic model, time dimension should contain
         time-delta objects
@@ -82,18 +81,23 @@ class PrognosticModel(Protocol):
         pass
 
     def output_coords(self, input_coords: CoordSystem) -> CoordSystem:
-        """Output coordinate system of the prognostic model
+        """Output coordinate system of the prognostic model give an input coordinate
+        system.
 
         Parameters
         ----------
         input_coords : CoordSystem
             Input coordinate system to transform into output_coords
-            by default None, will use self.input_coords.
 
         Returns
         -------
         CoordSystem
             Coordinate system dictionary
+
+        Raises
+        ------
+        ValueError
+            If input_coords are not valid
         """
         pass